+++ changelog.md
@@ -9,6 +9,8 @@
 
 ## Bug fixes
 
+* Fixed a bug with `from_gds` where graphs with different relationship types would fail if they had different properties.
+
 
 ## Improvements